+++ database/migrations/2017_11_01_145855_create_auctions_table.php
@@ -18,12 +18,12 @@ class CreateAuctionsTable extends Migration
             $table->integer('property_id')->unsigned();
             $table->integer('invitation_id')->unsigned();
             $table->string('additional_note', 255);
-            $table->float('min_price');
+            $table->double('min_price',13,2);
             $table->string('min_word');
             $table->string('organizer');
             $table->string('organizer_account');
             $table->string('account_name');
-            $table->float('price');
+            $table->double('price',15,2);
             $table->integer('percent');
             $table->string('word');
             $table->timestamps();
